tinyVIT training and classifying
Had to troubleshoot around some errors in my environment with loading the classes and the dataset, the bottom half of the code's the training and classifying portion of it while the upper half generally deals with setting up the dataset. It's probably messy to go through as I tried an assortment of fixes and didnt change or trim it after it worked.